// helpers/signup-enums.ts
export enum Days {
  Min = 1,
  Max = 31,
}

export enum Months {
  January = "January",
  February = "February",
  March = "March",
  April = "April",
  May = "May",
  June = "June",
  July = "July",
  August = "August",
  September = "September",
  October = "October",
  November = "November",
  December = "December",
}

export enum Years {
  Min = 1900,
  Max = 2021,
}

export enum Countries {
  India = "India",
  UnitedStates = "United States",
  Canada = "Canada",
  Australia = "Australia",
  Israel = "Israel",
  NewZealand = "New Zealand",
  Singapore = "Singapore",
}